+++ test/wal-workspace-test.el
@@ -44,4 +44,17 @@
 
     (was-called-with wal/project-command '(install "Install project: "))))
 
+(ert-deftest test-wal/project-find-rg ()
+  (with-mock ((rg-read-pattern . #'wal/rt)
+              (rg-read-files . #'wal/rt)
+              (project-root . (lambda (&rest _) "/tmp/project"))
+              project-current
+              rg-run)
+
+    (defvar rg-command-line-flags-function)
+    (let ((rg-command-line-flags-function #'wal/rt))
+      (call-interactively 'wal/project-find-rg))
+
+    (was-called-with rg-run (list 'testing 'testing "/tmp/project" nil nil 'testing))))
+
 ;;; wal-workspace-test.el ends here